Another WIP Commit

+++ Ryu64.MIPS/Interpreter/Registers.cs
@@ -9,9 +9,9 @@ namespace Ryu64.MIPS
         public class R4300
         {
             public static long[] Reg = new long[32];
-            public static long HI;
-            public static long LO;
-            public static ulong PC;
+            public static long  HI;
+            public static long  LO;
+            public static uint  PC;
         }
 
         public class COP0
